+++ app/remind-birthday_test.go
@@ -23,11 +23,11 @@ func Test_remind_birthdays_successful(t *testing.T) {
 
 	recipients := []string{"spe@mail.com", "wsh@prov.fr"}
 
-	contactToRemind := remind.ContactBirthday{Name: "John Bar", BirthDate: testdata.BirthDate(1986, time.August, 31), Age: 32}
+	contactToRemind := contact.Contact{Name: "John Bar", BirthDate: testdata.BirthDate(1986, time.August, 31)}
 	reminder := remind.Reminder{CurrentDate: testdata.LocalDate(2018, time.August, 30), NbDaysBeforeBDay: 1}
 
 	contactProvider.On("GetContacts").Return(con, nil)
-	emailContacts := email.Contacts{Contacts: []remind.ContactBirthday{contactToRemind}, RemindDate: testdata.LocalDate(2018, time.August, 31)}
+	emailContacts := email.Contacts{Contacts: []contact.Contact{contactToRemind}, RemindDate: testdata.LocalDate(2018, time.August, 31)}
 	smtp.On("Send", emailContacts, recipients).Return(nil)
 
 	err := BirthdayPal{}.Exec(contactProvider, smtp, reminder, recipients)
@@ -63,9 +63,9 @@ func Test_remind_birthdays_fail_during_mail_sending(t *testing.T) {
 
 	recipients := []string{"spe@mail.com", "wsh@prov.fr"}
 
-	contactToRemind := remind.ContactBirthday{Name: "John Bar", BirthDate: testdata.BirthDate(1986, time.August, 31), Age: 32}
+	contactToRemind := contact.Contact{Name: "John Bar", BirthDate: testdata.BirthDate(1986, time.August, 31)}
 	reminder := remind.Reminder{CurrentDate: testdata.LocalDate(2018, time.August, 30), NbDaysBeforeBDay: 1}
-	emailContacts := email.Contacts{Contacts: []remind.ContactBirthday{contactToRemind}, RemindDate: testdata.LocalDate(2018, time.August, 31)}
+	emailContacts := email.Contacts{Contacts: []contact.Contact{contactToRemind}, RemindDate: testdata.LocalDate(2018, time.August, 31)}
 
 	contactProvider.On("GetContacts").Return(con, nil)
 	smtp.On("Send", emailContacts, recipients).Return(errors.New("wow"))
