+++ email/sender_test.go
@@ -1,7 +1,7 @@
 package email
 
 import (
-	birthday2 "github.com/s-petit/birthday-pal/remind"
+	"github.com/s-petit/birthday-pal/contact"
 	"github.com/s-petit/birthday-pal/testdata"
 	"github.com/stretchr/testify/assert"
 	"testing"
@@ -20,7 +20,7 @@ func Test_send(t *testing.T) {
 	birthday := time.Date(2016, time.August, 22, 0, 0, 0, 0, time.UTC)
 	sender := SMTPClient{Host: "localhost", Port: 2525}
 	c := Contacts{
-		Contacts: []birthday2.ContactBirthday{{Name: "ttf", BirthDate: birthday, Age: 34}}, RemindDate: birthday,
+		Contacts: []contact.Contact{{Name: "ttf", BirthDate: birthday}}, RemindDate: birthday,
 	}
 	e := sender.Send(c, []string{"recipient@test", "recipient2@test"})
 	assert.NoError(t, e)
@@ -33,7 +33,7 @@ func Test_no_send_when_contact_list_empty(t *testing.T) {
 	birthday := time.Date(2016, time.August, 22, 0, 0, 0, 0, time.UTC)
 	sender := SMTPClient{Host: "localhost", Port: 2525}
 	c := Contacts{
-		Contacts: []birthday2.ContactBirthday{}, RemindDate: birthday,
+		Contacts: []contact.Contact{}, RemindDate: birthday,
 	}
 	e := sender.Send(c, []string{"recipient@test", "recipient2@test"})
 	assert.NoError(t, e)
@@ -46,7 +46,7 @@ func Test_send_error(t *testing.T) {
 	birthday := time.Date(2016, time.August, 22, 0, 0, 0, 0, time.UTC)
 	sender := SMTPClient{Host: "localhost", Port: 2525}
 	c := Contacts{
-		Contacts: []birthday2.ContactBirthday{{Name: "ttf", BirthDate: birthday, Age: 34}}, RemindDate: birthday,
+		Contacts: []contact.Contact{{Name: "ttf", BirthDate: birthday}}, RemindDate: birthday,
 	}
 	e := sender.Send(c, []string{"recipient@test2"})
 	assert.EqualError(t, e, "454 4.1.1 Error: Relay access denied: test2")
